Sound repeat+broadcast unity (#12)
* Sound looping + Broadcast unity

Sound looping was lost in newer versions. This has now been restored so that it doesn't just play a single sound. This only happens for broadcasts that have a custom sound so it should not interfere with tannoy.

Temple service was using the same sound as kitchen. This was wrong as it 'taints' the kitchen sound. As such Temple now has a similar sound and style as ThankYou & HydrationReminder. Temple service now also resets itself after 10 minutes.

Lunch & Dinner have been adjusted to share the same style.

Medical has been adjusted to share a similar style to dishes request.

Blood has been adjusted to not blink the entire notice but only the red bars (as originally intended)

* portal button on admin

Show portal button on admin screen without elevated rights